refactor(split-exepnse): change split expense repo to post auth
