+++ app/src/main/java/com/soul/taskbreeze/core/di/RepositoryModule.kt
@@ -2,8 +2,8 @@ package com.soul.taskbreeze.core.di
 
 import com.soul.taskbreeze.auth.login.data.repository.LoginRepository
 import com.soul.taskbreeze.auth.login.domain.repository.LoginRepositoryImpl
-import com.soul.taskbreeze.auth.split_expense.data.repository.SplitExpenseRepositoryImpl
-import com.soul.taskbreeze.auth.split_expense.domain.repository.SplitExpenseRepository
+import com.soul.taskbreeze.post_auth.split_expense.data.repository.SplitExpenseRepositoryImpl
+import com.soul.taskbreeze.post_auth.split_expense.domain.repository.SplitExpenseRepository
 import com.soul.taskbreeze.pre_auth.pre_auth_loading.data.repository.PreAuthRepositoryImpl
 import com.soul.taskbreeze.pre_auth.pre_auth_loading.domain.repository.PreAuthRepository
 import dagger.Binds
